Kept main file cleaner, broke out DB
