+++ .gitignore
@@ -1,2 +1,4 @@
 target/
 .DS_Store
+.idea
+e
