+++ src/main/java/com/controller/Main.java
@@ -1,0 +1,16 @@
+package com.controller;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+/**
+ * Initialize the web-app
+ *
+ * Read https://spring.io/guides/gs/spring-boot/
+ */
+@SpringBootApplication
+public class Main {
+
+    public static void main(String[] args) throws Exception {
+        SpringApplication.run(Main.class, args);
+    }
+}
