+++ src/main/resources/banner.txt
@@ -1,8 +1,8 @@
-  _    _                _
- | |  | |              | |
- | |__| | ___ _ __ ___ | | ___   _
- |  __  |/ _ \ '__/ _ \| |/ / | | |
- | |  | |  __/ | | (_) |   <| |_| |
- |_|  |_|\___|_|  \___/|_|\_\\__,_|
+_________        .__  .__                        
+\_   ___ \  ____ |  | |  | _____  ______ ______  
+/    \  \/ /  _ \|  | |  | \__  \ \____ \\____ \ 
+\     \___(  <_> )  |_|  |__/ __ \|  |_> >  |_> >
+ \______  /\____/|____/____(____  /   __/|   __/ 
+        \/                      \/|__|   |__|    
 
 :: Built with Spring Boot :: ${spring-boot.version}
