+++ src/main/resources/templates/error.html
@@ -2,8 +2,8 @@
 
 <html xmlns:th="http://www.thymeleaf.org" th:replace="~{fragments/layout :: layout (~{::body},'error')}">
 
-  <body>
-    <p th:text="${message}">Exception message</p>
-  </body>
+<body>
+<p th:text="${message}">Exception message</p>
+</body>
 
 </html>
