--- src/main/resources/templates/hello.html
@@ -1,8 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org" th:replace="~{fragments/layout :: layout (~{::body},'hello')}">
-<body>
-  <div class="container">
-    <p th:text="${science}"/>
-  </div>
-</body>
-</html>
